pyproject.toml
[tool.poetry]
name = "lesson-33"
version = "0.1.0"
description = ""
authors = ["shonsky <shon.kzidenov@gmail.com>"]
readme = "README.md"
packages = [{include = "lesson_33"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
python-dotenv = "^1.0.0"
pymongo = {extras = ["srv"], version = "^4.3.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
